Moved tagging file into stand-alone class
